+++ lib/stripe/subscriptions/subscription_item.ex
@@ -11,6 +11,7 @@ defmodule Stripe.SubscriptionItem do
   @type t :: %__MODULE__{
           id: Stripe.id(),
           object: String.t(),
+          billing_thresholds: Stripe.Types.subscription_billing_thresholds() | nil,
           created: Stripe.timestamp(),
           deleted: boolean | nil,
           metadata: Stripe.Types.metadata(),
@@ -22,6 +23,7 @@ defmodule Stripe.SubscriptionItem do
   defstruct [
     :id,
     :object,
+    :billing_thresholds,
     :created,
     :deleted,
     :metadata,